moved math functions to mathfunctions
